requirements.txt
adal==1.2.7
aiofiles==23.1.0
aiohttp==3.8.4
aiokafka==0.8.1
aiosignal==1.3.1
alembic==1.9.4
altair==4.2.2
ansi2html==1.8.0
antlr4-python3-runtime==4.11.1
anyio==3.6.2
appdirs==1.4.4
argcomplete==2.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
autoviz==0.1.58
azure-common==1.1.28
azure-core==1.26.3
azure-graphrbac==0.61.1
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.2.2
azure-mgmt-resource==22.0.0
azure-mgmt-storage==21.0.0
azure-storage-blob==12.14.1
azureml-core==1.52.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
beautifulsoup4==4.11.2
black==23.7.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
bokeh==3.0.3
Boruta==0.3
boto3==1.26.74
botocore==1.29.74
branca==0.6.0
Brotli==1.0.9
cachetools==5.3.0
catalogue==1.0.2
catboost==1.1.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
chart-studio==1.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloud-init==23.3.3
cloudpickle==2.2.1
cmaes==0.9.1
colorama==0.4.3
colorcet==3.0.1
colorlog==6.7.0
colorlover==0.3.0
colour==0.1.5
comm==0.1.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.0.7
cryptography==2.8
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
dash==2.8.1
dash-auth==1.4.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.4
db-dtypes==1.1.1
dbus-python==1.2.16
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
docker==6.0.1
dtreeviz==2.1.4
ec2-hibinit-agent==1.0.0
emoji==2.2.0
entrypoints==0.3
evidently==0.2.5
exceptiongroup==1.1.0
executing==1.2.0
explainerdashboard==0.4.2
fairlearn==0.8.0
fastapi==0.92.0
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.9.0
Fiona==1.9.0
Flask==2.2.3
Flask-Compress==1.13
Flask-SeaSurf==1.1.1
flask-simplelogin==0.1.1
Flask-WTF==0.15.1
fonttools==4.38.0
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.1.0
fugue==0.8.1
fugue-sql-antlr==0.1.5
funcy==1.18
future==0.18.3
gensim==3.8.3
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.16.1
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gradio==3.18.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==20.1.0
h11==0.14.0
hibagent==1.0.1
holoviews==1.15.4
htmlmin==0.1.12
httpcore==0.16.3
httplib2==0.14.0
httpx==0.23.3
humanfriendly==10.0
hvplot==0.8.2
hyperlink==19.0.0
hyperopt==0.2.7
idna==2.8
ImageHash==4.3.1
imageio==2.25.1
imbalanced-learn==0.7.0
importlib-metadata==5.2.0
importlib-resources==5.12.0
incremental==16.10.1
iniconfig==2.0.0
interpret==0.2.4
interpret-core==0.3.0
ipykernel==6.16.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.22
jsonpickle==3.0.1
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter_client==7.4.7
jupyter-console==6.5.1
jupyter_core==5.2.0
jupyter-dash==0.4.2
jupyter-events==0.6.3
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kafka-python==2.0.2
keyring==18.0.1
kiwisolver==1.4.4
kmodes==0.12.2
knack==0.10.1
kubernetes==26.1.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightgbm==3.3.5
lime==0.2.0.1
linkify-it-py==1.0.3
llvmlite==0.37.0
m2cgen==0.10.0
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.4
mdurl==0.1.2
mistune==2.0.5
mlflow==2.1.1
mlserver==1.3.5
mlserver-mlflow==1.3.5
mlxtend==0.21.0
more-itertools==4.2.0
msal==1.22.0
msal-extensions==1.0.0
msgpack==1.0.4
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.4
multimethod==1.9.1
multiprocess==0.70.14
munch==2.5.0
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.54.1
numexpr==2.8.4
numpy==1.23.1
oauthlib==3.1.0
optuna==3.1.0
orjson==3.8.6
oyaml==1.0
packaging==21.3
pandas==1.4.3
pandas-gbq==0.17.9
pandas-profiling==3.6.6
pandocfilters==1.5.0
panel==0.12.7
param==1.12.3
paramiko==3.2.0
parso==0.8.3
pathspec==0.11.1
patsy==0.5.3
pexpect==4.6.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
pip==23.3.2
pkginfo==1.9.6
plac==1.1.3
platformdirs==2.6.2
plotly==5.13.0
pluggy==1.0.0
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.16.0
prometheus-flask-exporter==0.22.4
prompt-toolkit==3.0.32
proto-plus==1.22.3
protobuf==4.22.0
psutil==5.9.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-grpc-prometheus==0.7.0
py4j==0.10.9.7
pyamg==4.2.3
pyaml==21.10.1
pyarrow==9.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pycryptodome==3.17
pyct==0.5.0
pydantic==1.10.5
pydata-google-auth==1.8.2
pydub==0.25.1
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pyLDAvis==3.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynndescent==0.5.8
pyod==1.0.7
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyproj==3.4.1
pyrsistent==0.15.5
pyserial==3.4
pysftp==0.2.9
pytest==7.2.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-json-logger==2.0.6
python-multipart==0.0.5
python-rapidjson==1.10
pytz==2022.7.1
pyviz-comms==2.2.1
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==24.0.1
qpd==0.4.0
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
ray==2.2.0
regex==2022.10.31
requests==2.31.0
requests-auth-aws-sigv4==0.7
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.0
SALib==1.4.7
scikit-image==0.19.3
scikit-learn==0.23.2
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.12.2
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
setuptools==45.2.0
shap==0.41.0
simplejson==3.16.0
six==1.14.0
skope-rules==1.0.1
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sos==4.5.6
soupsieve==2.4
spacy==2.3.9
SQLAlchemy==1.4.46
sqlglot==11.1.3
sqlparse==0.4.3
srsly==1.0.6
ssh-import-id==5.10
stack-data==0.6.1
starlette==0.25.0
starlette-exporter==0.16.0
statsmodels==0.13.5
systemd-python==234
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.1
tensorboardX==2.6
terminado==0.17.1
textblob==0.17.1
thinc==7.4.6
threadpoolctl==3.1.0
tifffile==2023.2.3
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
treeinterpreter==0.2.3
triad==0.8.2
tritonclient==2.35.0
tune-sklearn==0.4.5
Twisted==18.9.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
ua-parser==0.16.1
ubuntu-advantage-tools==8001
uc-micro-py==1.0.1
ufw==0.36
umap-learn==0.5.3
unattended-upgrades==0.1
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
visions==0.7.5
wadllib==1.3.3
waitress==2.1.2
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
wheel==0.34.2
widgetsnbextension==4.0.5
wordcloud==1.8.2.2
WTForms==3.0.1
xgboost==1.7.4
xlrd==2.0.1
xyzservices==2023.2.0
yarl==1.8.2
ydata-profiling==4.0.0
yellowbrick==1.5
zipp==3.14.0
zope.event==4.6
zope.interface==4.7.1
